Added Lazerian
Credit to Grab for most of the aesthetic work on this one. Unfortunately neither of us can find our .map files from before, so if any changes need to be made (such as missing textures), someone else will have to put in the labor. 

Map is intended for 4v4 Capture the Flag PUG gameplay.
Layout inspirations came primarily from Dosuun and CTF4.